created hook for chart filter
