+++ shesha-reactjs/src/designer-components/charts/bar.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState, useMemo, useRef } from 'react';
+import React from 'react';
 import { ConfigurableFormItem } from '@/components';
 import { validateConfigurableComponentSettings } from '@/formDesignerUtils';
 import { IToolboxComponent } from '@/interfaces';
@@ -12,11 +12,10 @@ import ChartControlURL from './chartControlURL';
 import ChartDataProvider from '@/providers/chartData';
 import ChartControl from './chartControl';
 import { migratePrevStyles } from '../_common-migrations/migrateStyles';
-import { evaluateDynamicFiltersSync } from '@/utils';
-import { useAvailableConstantsData, useDataContextManager, useMetadataDispatcher, IModelMetadata } from '@/index';
 import { useShaFormDataUpdate } from '@/providers/form/providers/shaFormProvider';
 import useStyles from './styles';
 import ChartLoader from './components/chartLoader';
+import { useChartFilters } from './hooks/useChartFilters';
 
 const BarChartComponent: IToolboxComponent<IChartProps> = {
   type: 'barChart',
@@ -26,79 +25,8 @@ const BarChartComponent: IToolboxComponent<IChartProps> = {
   icon: <BarChartOutlined />,
   Factory: ({ model }) => {
     useShaFormDataUpdate();
-    const allAvailableData = useAvailableConstantsData();
-    const dataContextManager = useDataContextManager();
-    const { getMetadata } = useMetadataDispatcher();
-    const [stateEvaluatedFilters, setStateEvaluatedFilters] = useState<string>('');
-    const [metaData, setMetaData] = useState<IModelMetadata>(undefined);
-    const [filtersReady, setFiltersReady] = useState<boolean>(false);
-    const [filterError, setFilterError] = useState<string | undefined>(undefined);
+    const { stateEvaluatedFilters, filtersReady, filterError } = useChartFilters(model);
     const { cx, styles } = useStyles();
-
-    // Use refs to track current filter state and prevent race conditions
-    const filtersRef = useRef<string>('');
-    const filtersReadyRef = useRef<boolean>(false);
-
-    useEffect(() => {
-      getMetadata({ modelType: model.entityType, dataType: 'entity' })
-        .then(setMetaData)
-        .catch(error => {
-          console.error('Error getting entity metadata:', error);
-          setFilterError('Error getting entity metadata');
-        });
-    }, [model.entityType]);
-
-    // Memoize the data context values to prevent unnecessary re-renders
-    const pageContext = useMemo(() => dataContextManager?.getPageContext(), [dataContextManager]);
-    const contextsData = useMemo(() => dataContextManager?.getDataContextsData(), [dataContextManager]);
-
-    useEffect(() => {
-      if (!model.filters) {
-        filtersRef.current = '';
-        filtersReadyRef.current = true;
-        setStateEvaluatedFilters('');
-        setFiltersReady(true);
-        setFilterError(undefined);
-        return;
-      }
-
-      // Check if we have all required data for filter evaluation
-      if (!metaData?.properties) {
-        console.warn('Waiting for metadata to evaluate filters');
-        return;
-      }
-
-      const match = [
-        { match: 'data', data: allAvailableData.form?.data },
-        { match: 'globalState', data: allAvailableData.globalState },
-        { match: 'pageContext', data: pageContext },
-        contextsData ? { match: 'contexts', data: contextsData } : null
-      ].filter(Boolean);
-
-      try {
-        const response = evaluateDynamicFiltersSync(
-          [{ expression: model.filters } as any],
-          match,
-          metaData?.properties
-        );
-
-        const strFilters = JSON.stringify(response[0]?.expression || '');
-
-        // Update both ref and state atomically
-        filtersRef.current = strFilters;
-        filtersReadyRef.current = true;
-        setStateEvaluatedFilters(strFilters);
-        setFiltersReady(true);
-        setFilterError(undefined);
-      } catch (error) {
-        console.error('Error evaluating filters:', error);
-        filtersRef.current = '';
-        filtersReadyRef.current = false;
-        setStateEvaluatedFilters('');
-        setFiltersReady(false);
-        setFilterError(error instanceof Error ? error.message : 'Error evaluating filters');
-      }
-    }, [metaData?.properties, model.filters, allAvailableData.form?.data, allAvailableData.globalState, pageContext, contextsData]);
 
     const {
       dimensionsStyles,
@@ -140,7 +68,7 @@ const BarChartComponent: IToolboxComponent<IChartProps> = {
       return (
         <ConfigurableFormItem model={model}>
           <div className={cx(styles.loadingContainer)}>
-            <ChartLoader chartType={model.chartType} /> 
+            <ChartLoader chartType={model.chartType} />
             <div className={cx(styles.loadingText)}>Fetching data...</div>
           </div>
         </ConfigurableFormItem>
